% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.classIntervals}
\alias{logLik.classIntervals}
\title{Log-likelihood for classIntervals objects}
\usage{
\method{logLik}{classIntervals}(object, ...)
}
\arguments{
\item{object}{A classIntervals object}

\item{...}{Ignored.}
}
\value{
A `logLik` object (see `stats::logLik`).
}
\description{
Log-likelihood for classIntervals objects
}
\examples{
x <- classIntervals(rnorm(100), n=5, style="fisher")
logLik(x)
AIC(x) # By having a logLik method, AIC.default is used.
}
\references{
Lucien Birge, Yves Rozenholc.  How many bins should be put in a regular
histogram.  ESAIM: Probability and Statistics. 31 January 2006. 10:24-45.
url: https://www.esaim-ps.org/articles/ps/abs/2006/01/ps0322/ps0322.html.
doi:10.1051/ps:2006001

Laurie Davies, Ursula Gather, Dan Nordman, Henrike Weinert. A comparison of
automatic histogram constructions. ESAIM: Probability and Statistics. 	11
June 2009.  13:181-196. url:
https://www.esaim-ps.org/articles/ps/abs/2009/01/ps0721/ps0721.html
doi:10.1051/ps:2008005
}
